pyproject.toml
[tool.poetry]
name = "azulrl"
version = "0.1.0"
description = ""
authors = ["Stefan Wanckel <stefanwanckel@gmail.com>"]
readme = "README.md"
packages = [{include = "azul", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
strenum = "^0.4.15"
black = "^25.1.0"
isort = "^6.0.1"
pytest = "^8.4.0"
python-statemachine = {extras = ["diagrams"], version = "^2.5.0"}

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''
[tool.pytest.ini_options]
testpaths = ["tests/unit", "tests/integration"]
markers = [
    "unit: marks tests as unit tests (fast, isolated)",
    "integration: marks tests as integration tests (slower, with dependencies)"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
